==> Projects/project_3-master/code/01_get_data.ipynb <==
## **Notebook Contents**

- [Import Libraries](#importlibraries4)  
- [Data Gathering](#gatherdata4)
- [Future Steps](#steps4)  
- [Citations](#cites4)

<a name="importlibraries4"></a>
## **Import Libraries**

import requests
import time
import pandas as pd

<a name="gatherdata4"></a>
## **Data Gathering**

# Get Submission 
def submission_request_ml(**kwargs):
    
    # This does a get request to pull Submissions from the Machine Learning subreddit       #kwargs will be the params
    request = requests.get("https://api.pushshift.io/reddit/search/submission/?subreddit=MachineLearning", kwargs)
    # This gets us data from the pull request
    data = request.json()
    # This gets us the individual 'DATA' for each post --> add [0] to get the first post
    return data['data']

def submission_request_ai(**kwargs):
    
    # This does a get request to pull Submissions from the Machine Learning subreddit       #kwargs will be the params
    request = requests.get("https://api.pushshift.io/reddit/search/submission/?subreddit=artificial", kwargs)
    # This gets us data from the pull request
    data = request.json()
    # This gets us the individual 'DATA' for each post --> add [0] to get the first post
    return data['data']

subreddit_ml = []
title_ml = []
selftext_ml =[]

subreddit_ai = []
title_ai = []
selftext_ai =[]

# # Create a dataframe for ml
# pd.DataFrame(columns=[''])

before_ml = None
before_ai = None
iterations_times_25 = 2500  # This is the number of times it runs --> that number times 25 is how many rows we "should" get

# Looping through the iterations 
while iterations_times_25 != 0:
    time.sleep(1.5)
    ml_submissions = submission_request_ml(before = before_ml, sort='desc', sort_type='created_utc')
    ai_submissions = submission_request_ai(before = before_ai, sort='desc', sort_type='created_utc')

#   MACHINE LEARNING
    for xi in ml_submissions:
        try:
            before_ml = xi['created_utc']
            
            selftext_ml.append(xi['selftext'])
            subreddit_ml.append(xi['subreddit'])
            title_ml.append(xi['title'])
            
        except:
            pass

#   ARTIFICAL INTELLIGENCE    
    for n in ai_submissions:
        try:
            
            before_ai = n['created_utc']
            
            selftext_ai.append(n['selftext'])
            subreddit_ai.append(n['subreddit'])
            title_ai.append(n['title'])
            
        except:
            pass
        
    time.sleep(1.5)
    iterations_times_25 -= 1

print(len(subreddit_ml))
print(len(title_ml))
print(len(selftext_ml))

print(len(subreddit_ai))
print(len(title_ai))
print(len(selftext_ai))

data_ml = pd.DataFrame({'subreddit': subreddit_ml,
                        'title': title_ml,
                        'selftext': selftext_ml
                       })

data_ml.shape

data_ai = pd.DataFrame({'subreddit': subreddit_ai,
                        'title': title_ai,
                        'selftext': selftext_ai
                       })

data_ai.shape

# This will change based on how many rows we are able to get from reddit.
data_ml = data_ml[:31299]

data_ml.shape

data_ai.shape

data_ml.to_csv('./data/data_ml.csv', index=False)

data_ai.to_csv('./data/data_ai.csv', index=False)

<a name="steps4"></a>
## **Future Steps**

In the future I plan to make these functions able to scrape all of the 138,000 active subreddits and pull 2,000 posts from each subreddit. After this step I will have 276 million posts scraped and able to perform analysis on. 

<a name="cites4"></a>
## **Citations**

Credit to David Lee for assitance with these steps  

Crawling the submissions:  
- https://www.textjuicer.com/2019/07/crawling-all-submissions-from-a-subreddit/

Function for getting the submission:  
- https://pythonprogramming.altervista.org/collect-data-from-reddit/?doing_wp_cron=1597670992.0452320575714111328125

Countdown While loop for each iteration:  
- https://datatofish.com/while-loop-python/

Try and Except for 'selftext' problem:  
- https://stackoverflow.com/questions/38707513/ignoring-an-error-message-to-continue-with-the-loop-in-python
