==> Projects/project_3-master/code/02_baseline_score.ipynb <==
## **Notebook Contents**

- [Import Libraries](#importlibraries2)  
- [Import Dataframes](#importdataframes2)
- [Merge the Data](#mergedata2)  
- [X & Y](#xandy2)
- [Baseline Score](#baseline2)

<a name="importlibraries2"></a>
## **Import Libraries**

import pandas as pd

<a name="importdataframes2"></a>
## **Import Dataframes**

data_ai = pd.read_csv('../data/data_ai.csv')
data_ml = pd.read_csv('../data/data_ml.csv')

data_ai.head(1)

data_ml.head(1)

<a name="mergedata2"></a>
## **Merge the Data**

df = data_ai.append(data_ml).reset_index()

df.drop(columns='index',inplace=True)

df

**Let's see what a title might look like:**

df['title'][0]

**Let's see what a selftext might look like:**

df['selftext'][3228]

<a name="xandy2"></a>
## **X & Y**

X = df[['title']]
y = df['subreddit']

X.head(1)

X.shape

y.shape

<a name="baseline2"></a>
## **Baseline Score**

# This is the baseline
y.value_counts(normalize=True)

The baseline score is 50% accuracy because if we had to randomly pick a subreddit from a post we would have a 50% accuracy in picking either subreddit since our classes are evenly sampled.  

This is the score to beat!
